Initial commit.
At this stage, we parse GET requests, and try to serve the file requested, if it exists in this working dir. We give a 404 error if the file is not found. Other cool things are the fancy buffer printer to visualize CRLF confusion, and a workaround for the socket acting like its not closed: if we try and the socket is broken, just generate another.